1_ingesta/api_producer/requirements.txt
requests>=2.28.0
kafka-python>=2.0.2
